requirements.txt
django-debug-toolbar
django-redis
django-tastypie
ipython
